Restructure crate features
Resolves #184
