+++ benches/cipher_decrypt.rs
@@ -28,8 +28,10 @@ pub struct CipherDecrypt {
 
 impl CipherDecrypt {
     pub fn random(rng: &mut StdRng) -> Self {
-        let shared =
-            GENERATOR.to_niels().mul(&JubJubScalar::random(rng)).into();
+        let shared = GENERATOR
+            .to_niels()
+            .mul(&JubJubScalar::random(&mut *rng))
+            .into();
         let nonce = BlsScalar::random(&mut *rng);
         let message =
             [BlsScalar::random(&mut *rng), BlsScalar::random(&mut *rng)];
@@ -44,14 +46,11 @@ impl CipherDecrypt {
 }
 
 impl Circuit for CipherDecrypt {
-    fn circuit<C>(&self, composer: &mut C) -> Result<(), Error>
-    where
-        C: Composer,
-    {
+    fn circuit(&self, composer: &mut Composer) -> Result<(), Error> {
         let shared = composer.append_point(self.shared);
         let nonce = composer.append_witness(self.nonce);
 
-        let mut cipher_circuit = [C::ZERO; CIPHER_SIZE];
+        let mut cipher_circuit = [Composer::ZERO; CIPHER_SIZE];
         self.cipher
             .cipher()
             .iter()
